Changed aux-init to initialize variables better.  I think it handles just about all cases now.
Modified clean-dcls to handle simple-arrays.

Changed collect so that it handles types better by passing the correct
type to fragL.  This allows better optimization by the compiler (at
least for CMUCL).

Added code at the end so that the package is installed whenever it's
loaded.  You don't have to explicitly install the package anymore.
However, there's a bug:  It assumes you were originally in the USER
package.  This needs to be fixed.