Bug #115738:
Use remove-if-not instead of delete-if-not in delete-aux-if-not.  This
was causing CLISP to fail test 530.

(I'm not sure about this.  It seems there's some shared list structure
with CLISP that doesn't happen in CMUCL.  However, I think it's safe
to cons up a new list instead of destructively modifying the
original.)